refactor: 보호막 수량 컬럼 타입 조정

+++ backend/bom-bom-server/src/main/java/me/bombom/api/v1/reading/domain/ContinueReadingShield.java
@@ -32,7 +32,7 @@ public class ContinueReadingShield extends BaseEntity {
     @Column(nullable = false)
     private Long memberId;
 
-    @Column(nullable = false, columnDefinition = "SMALLINT")
+    @Column(nullable = false, columnDefinition = "TINYINT")
     private int remainingCount;
 
     @Builder
